This is a functioning main program as of 5/19/22
This main script accounted for the allocation of the time iteration step using hourly calculations but the dx values weren't scaled. Need to determine whether or not there is a memory allocation issue it we go back to seconds. Can make a function to handle what is dx and dx_1 if they want the aggregation of seconds, minutes, or hours. Successful run simulation occurred 19May22b simulations and 13May22e. (However this reflects those changes. Now focusing back on debugging and which time step should be include. This includes the read in and use of real data from HMS API request retrievals.